improve rtl handling in pagination and front
This closes issue #1513
It sets `icon` to the left arrow on the right button and vice-versa if the user language's direction is `rtl`, as we'd still expect the behavior of the buttons and pagination dots to reverse alongside the language.

Additionally, it styles the class `.front__drawings` to display in Left to Right no matter the direction of the user's language.